#190: 表決 done

+++ voter_guide/votes/views.py
@@ -7,6 +7,8 @@ from django.db import IntegrityError, transaction
 
 from .models import Votes, Councilors_Votes
 from councilors.models import CouncilorsDetail
+from candidates.models import Intent, Intent_Standpoints
+from candidates.forms import Intent_StandpointsForm
 from search.views import keyword_list, keyword_been_searched
 from standpoints.models import Standpoints, User_Standpoint
 from commontag.views import paginate
@@ -53,24 +55,44 @@ def votes(request, county):
 
 def vote(request, vote_id):
     vote = get_object_or_404(Votes.objects.select_related('sitting'), uid=vote_id)
+    intent = None
     if request.user.is_authenticated():
+        try:
+            intent = Intent.objects.get(user=request.user)
+        except:
+            instance = None
+        else:
+            try:
+                instance = Intent_Standpoints.objects.get(intent=intent, vote_id=vote_id)
+            except:
+                instance = None
+        if intent:
+            form = Intent_StandpointsForm(instance=instance)
         if request.POST:
             with transaction.atomic():
-                if request.POST.get('keyword', '').strip():
-                    standpoint_id = u'vote-%s-%s' % (vote_id, request.POST['keyword'].strip())
-                    standpoint, created = Standpoints.objects.get_or_create(uid=standpoint_id, county=vote.sitting.county, title=request.POST['keyword'].strip(), vote_id=vote_id, user=request.user)
-                    if created:
-                        User_Standpoint.objects.create(standpoint_id=standpoint_id, user=request.user)
-                        Standpoints.objects.filter(uid=standpoint_id).update(pro=F('pro') + 1)
-                        tag_create_achievement(request.user)
-                        tag_pro_achievement(standpoint_id)
-                elif request.POST.get('pro'):
-                    User_Standpoint.objects.create(standpoint_id=request.POST['pro'], user=request.user)
-                    Standpoints.objects.filter(uid=request.POST['pro']).update(pro=F('pro') + 1)
-                    tag_pro_achievement(request.POST['pro'])
-                elif request.POST.get('against'):
-                    User_Standpoint.objects.get(standpoint_id=request.POST['against'], user=request.user).delete()
-                    Standpoints.objects.filter(uid=request.POST['against']).update(pro=F('pro') - 1)
+                if request.POST.has_key('intent') and intent:
+                    form = Intent_StandpointsForm(request.POST, instance=instance)
+                    if form.has_changed() and form.is_valid():
+                        intent_sp = form.save(commit=False)
+                        intent_sp.intent = intent
+                        intent_sp.vote = vote
+                        intent_sp.save()
+                else:
+                    if request.POST.get('keyword', '').strip():
+                        standpoint_id = u'vote-%s-%s' % (vote_id, request.POST['keyword'].strip())
+                        standpoint, created = Standpoints.objects.get_or_create(uid=standpoint_id, county=vote.sitting.county, title=request.POST['keyword'].strip(), vote_id=vote_id, user=request.user)
+                        if created:
+                            User_Standpoint.objects.create(standpoint_id=standpoint_id, user=request.user)
+                            Standpoints.objects.filter(uid=standpoint_id).update(pro=F('pro') + 1)
+                            tag_create_achievement(request.user)
+                            tag_pro_achievement(standpoint_id)
+                    elif request.POST.get('pro'):
+                        User_Standpoint.objects.create(standpoint_id=request.POST['pro'], user=request.user)
+                        Standpoints.objects.filter(uid=request.POST['pro']).update(pro=F('pro') + 1)
+                        tag_pro_achievement(request.POST['pro'])
+                    elif request.POST.get('against'):
+                        User_Standpoint.objects.get(standpoint_id=request.POST['against'], user=request.user).delete()
+                        Standpoints.objects.filter(uid=request.POST['against']).update(pro=F('pro') - 1)
 
     standpoints_of_vote = Standpoints.objects.filter(vote_id=vote_id)\
                                              .order_by('-pro')
@@ -78,4 +100,4 @@ def vote(request, vote_id):
         standpoints_of_vote = standpoints_of_vote.extra(select={
             'have_voted': "SELECT true FROM standpoints_user_standpoint su WHERE su.standpoint_id = standpoints_standpoints.uid AND su.user_id = %s" % request.user.id,
         },)
-    return render(request,'votes/vote.html', {'vote': vote, 'standpoints_of_vote': standpoints_of_vote})
+    return render(request,'votes/vote.html', {'vote': vote, 'standpoints_of_vote': standpoints_of_vote, 'intent': intent, 'form': intent and form})
